// src/component-library-contentsdk/consultant-finder/Locations/Practice/Practice.types.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export interface PracticeProps {
  name?: string;
  street1: string;
  street2: string;
  city: string;
  postcode: string;
  lat: any;
  long: any;
  slug: string;
  workingOpeningHours: any;
  facilityURL: string | null;
  viewOnGoogleMapsText: string;
}